Make test agreement data safe for Windows
The `jq` call seems to be inserting a carriage return (\r) at the end
of the creation of the constants AGREEMENT_TRIAL_RDV and
AGREEMENT_DRAFT_RDV, which is then interpolated into Json agreement data.
This may be because the curl query failed and the result is null, as
other values processed through `jq` don't have similar issues. This is
likely a bug in `jq`, as the carriage return character doesn't normally
appear without a line feed (\n), suggesting the line feed is being
cleaned from the output.